refactor(VideoPlayer): RichVideoPlayer を VideoPlayer にリネーム
- RichVideoPlayer → VideoPlayer に名称変更
- RichVideoPlayerProps → VideoPlayerProps に型名変更
- ディレクトリ名を src/components/VideoPlayer に変更
- バージョンを 0.20.0 に更新

BREAKING CHANGE: RichVideoPlayer は VideoPlayer に名称変更されました
