+++ src/components/VideoPlayer/ControlPanel.tsx
