+++ src/components/VideoPlayer/ProgressBar.tsx
